Use stdint.h for integer types

+++ experiment.c
@@ -1,39 +1,39 @@
-#include <stdio.h>
+#include <stdint.h>
 
 #include "experiment.h"
 
 #define MAX(a, b) ((a) > (b) ? (a) : (b))
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
 
-double colour_similarity(SDL_Colour a, SDL_Colour b)
+float colour_similarity(SDL_Colour a, SDL_Colour b)
 {
-    double total_similarity = 0;
+    float total_similarity = 0;
 
 
     if (a.r != 0 || b.r != 0)
-        total_similarity += MIN((double) a.r, (double) (b.r)) / MAX((double) a.r, (double) b.r);
+        total_similarity += MIN((float) a.r, (float) (b.r)) / MAX((float) a.r, (float) b.r);
     else
-        total_similarity += 1 - (MAX((double) a.r, (double) b.r) / 255);
+        total_similarity += 1 - (MAX((float) a.r, (float) b.r) / 255);
 
     if (a.g != 0 || b.g != 0)
-        total_similarity += MIN((double) a.g, (double) (b.g)) / MAX((double) a.g, (double) b.g);
+        total_similarity += MIN((float) a.g, (float) (b.g)) / MAX((float) a.g, (float) b.g);
     else
-        total_similarity += 1 - (MAX((double) a.g, (double) b.g) / 255);
+        total_similarity += 1 - (MAX((float) a.g, (float) b.g) / 255);
 
     if (a.b != 0 || b.b != 0)
-        total_similarity += MIN((double) a.b, (double) (b.b)) / MAX((double) a.b, (double) b.b);
+        total_similarity += MIN((float) a.b, (float) (b.b)) / MAX((float) a.b, (float) b.b);
     else
-        total_similarity += 1 - (MAX((double) a.b, (double) b.b) / 255);
+        total_similarity += 1 - (MAX((float) a.b, (float) b.b) / 255);
 
     return total_similarity / 3.0f;
 }
 
-double compare_canvasses(Canvas *A, Canvas *B)
+float compare_canvasses(Canvas *A, Canvas *B)
 {
     if ((A->w != B->w) || (A->h != B->h))
         return -1;
 
-    double similarity = 0;
+    float similarity = 0;
     uint32_t pixel_num = A->w * A->h;
 
     for (int i = 0; i < A->w; ++i)
@@ -43,7 +43,7 @@ double compare_canvasses(Canvas *A, Canvas *B)
                 A->data[i][j].b == 0 && B->data[i][j].b == 0)
                 --pixel_num;
             else
-                similarity += colour_similarity(A->data[i][j], B->data[i][j]) / (double) pixel_num;
+                similarity += colour_similarity(A->data[i][j], B->data[i][j]) / (float) pixel_num;
 
     return similarity;
 }
